backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pandas==2.1.4
numpy==1.26.3
xgboost==2.0.3
scikit-learn==1.4.0
joblib==1.3.2
requests==2.31.0
openpyxl==3.1.2
python-multipart==0.0.6
pytest==7.4.3
httpx==0.26.0
supabase==2.3.4
pdfplumber==0.11.0
pytesseract==0.3.10
python-docx==1.1.0
Pillow==10.2.0
faster-whisper==0.10.0
python-magic==0.4.27
pytz==2024.1
yfinance==0.2.36
